Vending Machine Updates Saturday

+++ 19_Capstone/java/src/main/java/com/techelevator/Candy.java
@@ -1,8 +1,10 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Candy extends Product {
 
-	public Candy(String productName, Double productPrice) {
+	public Candy(String productName, BigDecimal productPrice) {
 		super(productName, productPrice);
 	}
 
@@ -10,3 +12,4 @@ public class Candy extends Product {
 		return "Munch Munch, Yum!";
 	}
 }
+
