+++ 19_Capstone/java/src/main/java/com/techelevator/Chip.java
@@ -1,8 +1,10 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Chip extends Product {
 
-	public Chip(String productName, Double productPrice) {
+	public Chip(String productName, BigDecimal productPrice) {
 		super(productName, productPrice);
 	}
