+++ 19_Capstone/java/src/main/java/com/techelevator/Drink.java
@@ -1,8 +1,10 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Drink extends Product {
 
-	public Drink(String productName, Double productPrice) {
+	public Drink(String productName, BigDecimal productPrice) {
 		super(productName, productPrice);
 	}
