+++ 19_Capstone/java/src/main/java/com/techelevator/Gum.java
@@ -1,8 +1,10 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Gum extends Product {
 
-	public Gum(String productName, Double productPrice) {
+	public Gum(String productName, BigDecimal productPrice) {
 		super(productName, productPrice);
 	}
